Create README.md
This is my first website which I am writing for my friend. It is based on HTML, CSS and little JavaScript. The website is for a Personal Trainer and it is Responsive, which means that it is accessible on a mobile, tablet and laptop/ PC. The layout of the website changes depending on the device on which is viewed.